pyproject.toml
[project]
name = "halal-gpp"
version = "1.0.0"
description = "Halal Google Profile Protector"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "HalalGPP" }]
dependencies = [
  "PySide6>=6.7",
  "cryptography>=43",
  "argon2-cffi>=23",
  "zstandard>=0.22",
  "psutil>=5.9",
]

[project.scripts]
halal-gpp = "halal_gpp.app:main"

[project.optional-dependencies]
dev = [
  "nuitka>=2.4",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
